Add Changes

+++ OMS.API/OMS.Domain.Entities/API/Request/Organization/AddEditOrganizationOtherChargesRequest.cs
@@ -9,11 +9,11 @@ namespace OMS.Domain.Entities.API.Request.Organization
     public class AddEditOrganizationOtherChargesRequest
     {
         public short? OrganizationOtherChargeId {  get; set; }
-        public float? HandlingFees {  get; set; }
-        public float? BankWireFees { get; set; }
-        public float? CreditCardServiceFees { get; set; }
-        public float? ColdBoxFees { get; set; }
-        public float? ITNFees {  get; set; }
+        public decimal? HandlingFees {  get; set; }
+        public decimal? BankWireFees { get; set; }
+        public decimal? CreditCardServiceFees { get; set; }
+        public decimal? ColdBoxFees { get; set; }
+        public decimal? ITNFees {  get; set; }
 
     }
 }
